requirements.txt
Django>=4.2.0
django-import-export>=3.0.0
pandas>=2.0.0
reportlab>=4.0.0
Pillow>=10.0.0
requests>=2.31.0

# Production dependencies
gunicorn>=21.2.0
whitenoise>=6.6.0
dj-database-url>=2.1.0
psycopg2-binary>=2.9.9

